refactored convertFileToUnicode, new IT test

+++ src/test/java/org/dontexist/kb/integration/AbstractEpubConverterIT.java
@@ -19,8 +19,10 @@ abstract class AbstractEpubConverterIT {
 
     // --------------- HELPER METHODS -----------------
 
-    protected void verify(File inputfile, File actualOutputFile, File expectedOutputFile) throws IOException, Exception {
-        actualOutputFile.deleteOnExit();
+    protected void verify(File inputfile, File actualOutputFile, File expectedOutputFile, boolean deleteOutputFile) throws IOException, Exception {
+        if (deleteOutputFile) {
+            actualOutputFile.deleteOnExit();
+        }
         StringBuilder convertedFileAsString = unicodeConverterHelper.convertFileAsOneStringToUnicode(FileUtils.readFileToString(inputfile));
         FileUtils.writeStringToFile(actualOutputFile, convertedFileAsString.toString());
         FileAssert.assertEquals(expectedOutputFile, actualOutputFile);
